# backend/internal/handler/openapi.yaml
openapi: 3.0.3
info:
  title: BeefTV HTTP API
  version: "1.0.0"
  description: |
    路径前缀是 `/api`。前后端一起发布，不按路径做 API 版本。

    业务 JSON 信封为 `{ code, data, msg, reason }`。`code === 0` 表示成功；失败时 HTTP status 与
    `code` 都表达失败。`code` 默认等于 HTTP 状态；同一状态下的细分原因使用 `status*100+n`
    （例如 40301 配额、42901 限流）。机器判断请用 `reason`，不要解析 `msg`。

    TypeScript 调用面仍是 `web/src/services/api` 的 `http` 模块，不从本规范生成客户端。
servers:
  - url: /api
security:
  - sessionCookie: []
...
paths:
...
  /ops:
    get:
      tags: [agent-ops]
      summary: 列出调用者身份可用的操作描述（含 JSON Schema）
      description: |
        能力模式只由调用者身份决定：已登记客户端的模式来自服务端登记，请求体或查询参数都不能提升权限。
        只读客户端只会拿到 readOnly 操作；写操作在 /ops/{op} 上仍按登记权限强制拒绝。
      responses:
        "200":
          description: 信封 data.ops 为操作描述数组，data.readOnly 表示本次调用的能力模式
        "403":
          description: 非本机同源，或身份不足（缺少已登记客户端凭据与 owner 凭据）
  /ops/{op}:
    post:
      tags: [agent-ops]
      summary: 执行单个操作（CLI、MCP 与内置 pi 会话共用）
      description: |
        写操作必须提供 opId 作为幂等键：同一 opId 同参数回读原结果，不同参数返回冲突。
        写操作还必须在 params 里带上读取时的 expectedRevision：缺少或已过期都会被拒绝，
        不会覆盖别人刚写入的内容。字段落点由画布能力描述符声明（例如生成类节点的可编辑
        提示词落在 metadata.composerContent），操作层不自行拼文档形状。
      parameters:
        - name: op
          in: path
          required: true
          schema:
            type: string
          description: 操作 ID，例如 canvas.get、canvas.nodes.create、canvas.node.update
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                opId:
                  type: string
                  description: 写操作的幂等键；只读操作不接受该字段
                params:
                  type: object
                  description: 操作参数，形状以 GET /ops 返回的 JSON Schema 为准
      responses:
        "200":
          description: 信封 data 为操作结果；data.replayed=true 表示命中了幂等回放
        "400":
          description: 参数错误；reason=unsupported_field 表示该节点类型没有这个可编辑字段（整批不写）
        "409":
          description: 同一 opId 用于不同请求内容，或 revision 已过期（reason=stale_revision）
        "412":
          description: 缺少读取时的 expectedRevision
        "403":
          description: 只读客户端尝试写操作（reason=read_only_client），或身份不足
...
  /timeline/render-plan:
    post:
      tags: [tasks]
      summary: 编译时间线语义渲染计划
      description: |
        只读规划：根据时间线快照与不透明源元数据返回版本化语义计划。
        不排队、不落盘、不计费，不接受 ffmpeg 参数或本地路径。
      responses:
        "200":
          $ref: "#/components/responses/Envelope"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
  /timeline/renders:
    post:
      tags: [tasks]
      summary: 创建本机时间线渲染任务
      description: 接收时间线快照与输出选项（宽高、帧率、采样率、是否烧录字幕）。选项写入任务输入，由本机 ffmpeg 按语义计划执行。可选 clientOperationId：同一确认重复提交回读原任务。
      responses:
        "200":
          $ref: "#/components/responses/Envelope"
        "400":
          $ref: "#/components/responses/Error"
  /timeline/transcriptions:
    post:
      tags: [tasks]
      summary: 创建本机字幕转写任务
      description: 接收待转写素材编号和可选语言。只接受当前用户已上传完成的音视频。可选 clientOperationId：同一确认重复提交回读原任务。
      responses:
        "200":
          $ref: "#/components/responses/Envelope"
        "400":
          $ref: "#/components/responses/Error"
  /depth-captures:
    post:
      tags: [tasks]
      summary: 创建本机深度动作捕捉任务
      description: 接收待处理视频素材编号。只接受当前用户已上传完成的视频。可选 clientOperationId：同一确认重复提交回读原任务。
      responses:
        "200":
          $ref: "#/components/responses/Envelope"
        "400":
          $ref: "#/components/responses/Error"
  /tasks:
    post:
      tags: [tasks]
      summary: 创建生成任务
      responses:
        "200":
          $ref: "#/components/responses/Envelope"
        "400":
          $ref: "#/components/responses/Error"
...
components:
...
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: open_ai_canvas_session
  schemas:
    Envelope:
      type: object
      required: [code, msg]
      properties:
        code:
          type: integer
          description: 0 成功；失败时默认等于 HTTP 状态，或 status*100+n 的细分码
        data: {}
        msg:
          type: string
        reason:
          type: string
          description: 仅失败时出现。稳定机器可读原因。
    ErrorEnvelope:
      allOf:
        - $ref: "#/components/schemas/Envelope"
        - type: object
          properties:
            code:
              type: integer
              enum: [400, 401, 403, 404, 409, 429, 500, 502, 503, 504, 40301, 40901, 42901]
            data:
              nullable: true
            reason:
              type: string
              enum:
                - invalid_argument
                - unauthorized
                - forbidden
                - not_found
                - conflict
                - failed_precondition
                - quota_exceeded
                - rate_limited
                - unavailable
                - timeout
                - internal
                - local_storage_failed
                - bad_gateway
                - model_capability_not_supported
                - model_price_not_configured
                - model_route_unavailable
                - provider_request_failed
                - model_catalog_mismatch
                - invalid_model_selection
  responses:
    Envelope:
      description: 业务成功信封
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Envelope"
    Error:
      description: 业务失败信封
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"